projects/phenotype_library/requirements.txt
# Tested in Python 3.11.2 environment
-e ../../
numpy
pandas>=2, <3
PyYAML>=6, <7
fastparquet
lxml
python-dotenv
streamlit
matplotlib
datetime
plotly
